requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
Django==3.1.2
django-cors-headers==3.5.0
django-extensions==3.0.9
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.4
html5lib==1.1
psycopg2==2.8.6
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
webencodings==0.5.1